mainapp/requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
django-crispy-forms
django-bootstrap4
python-twitter
sqlalchemy
jsonlines
django-q